Do not insert priv_remove() when unneeded.
Modified the PrivRemoveInsert pass so that it does not insert
calls to priv_remove() in basic blocks that end with an
unreachable instruction or right before an llvm.trap() intrinsic
call.  Such basic blocks are about to terminate, and adding such
priv_remove() calls makes the bitcode more difficult to read when
debugging.