Fixed privilege propagation bug.
Enhanced the Privilege Use Summary Analysis so that it records
which functions are called from external code.

Fixed the Global Live Privilege Analysis so that it correctly
propagates live privileges from the external calling node in the
call graph (which represents external code) to functions called by
external code.

This fixes sshd.