Simplified drop computation code.
Simplified the code that computes which privileges need to be
removed (dropped) for each basic block.  The new code now skips
over basic blocks that don't need to remove any privileges
instead of adding information for them and then removing that
information from the data structures.

This may also fix some obscure bug in which the results differed
slightly between executions.  I cannot trigger the bug with this
change.