Added two helper function for debugging
I observed a very interesting but confusing fact: any function inserted
at the beginning of the main function of sshd would be called twice,
even though the code shows they should be only called once.

In my test, I inserted a function

void calledAtStart() {
    fprintf(stderr, "===JZ: calledAtStart() is called\n");
    fprintf(stderr, "===JZ: addBBLOICallCount = %lu\n", addBBLOICallCount);
}

at the very beignning of main. I checked the final binary executable of
sshd and confirmed that calledAtStart is only called at the beinning
of main. However, when I ran sshd in the foreground, I saw that
calledAtStart is called twice. The second call doesn't not immediately
follow the first call but after some debugging information printed out
from the sshd source code. Both of these two calls happen before
the first call to fork().

I also found that when the function is called the second time,
all the global variables have the same value as what they are when
the function is called the first time. For example, I have a counter addBBLOICallCount
to count the number of times function addBBLOI is called. This counter
is initialized to 0 when it's defined. Every time addBBLOI is called,
counter is increased by 1 and when it reaches a multiple of 20,000,
it is printed out. Before the calledAtStart() is called
the second time, the counter shows that addBBLOI has been called
at least 40,000 times. However, the second call of calledAtStart()
shows that addBBLOICallCount == 0. I also tested other global variables
and found that when the second call happens, no gloable vairables changes from
their original values.

This phenomenon is really confusing to me!